Moved event card to share components and removed unused functions in homeview
