+++ frontend/src/components/shared/EventCard.vue
